General refactoring of Animal class
The Cell object now keeps track of the animals on it - this is so other animals
can see each other through the cells in their sight radius

Revamped the Needs system. Needs have a rate of decline and a refresh rate,
animals pick the needs with the lowest level, each move depreciates the other needs by their rate of decline.
Animals on a tile that satisfy their objective need have that need appreciated by the refresh rate
Animals currently have 2 types of need - Terrain needs such as Sleep and Water, and RelationalNeeds such as Sex and Carnivorous eating
Terrain needs are satisfied by the tile the animal is on - Relational needs are satisfied by a set of criteria passed to the need